Testing on Linux

+++ include/qt_SnapViewer/s4SnapTableModel_L2Stats.h
@@ -106,19 +106,19 @@ namespace S4
             std::vector<QVariant> data;
             for (auto& key : _row_names){
                 switch(key){
-                    case dataType_t::recv_frame_cnt : data.emplace_back(stats.recv_frame_cnt); break;
-                    case dataType_t::recv_heartbeat_cnt : data.emplace_back(stats.recv_heartbeat_cnt); break;
-                    case dataType_t::recv_instrument_snap_cnt : data.emplace_back(stats.recv_instrument_snap_cnt); break;
-                    case dataType_t::recv_index_snap_cnt : data.emplace_back(stats.recv_index_snap_cnt); break;
-                    case dataType_t::recv_order_cnt : data.emplace_back(stats.recv_order_cnt); break;
-                    case dataType_t::recv_exec_cnt : data.emplace_back(stats.recv_exec_cnt); break;
-                    case dataType_t::recv_unknown_cnt : data.emplace_back(stats.recv_unknown_cnt); break;
+                    case dataType_t::recv_frame_cnt : data.emplace_back((long long)stats.recv_frame_cnt); break;
+                    case dataType_t::recv_heartbeat_cnt : data.emplace_back((long long)stats.recv_heartbeat_cnt); break;
+                    case dataType_t::recv_instrument_snap_cnt : data.emplace_back((long long)stats.recv_instrument_snap_cnt); break;
+                    case dataType_t::recv_index_snap_cnt : data.emplace_back((long long)stats.recv_index_snap_cnt); break;
+                    case dataType_t::recv_order_cnt : data.emplace_back((long long)stats.recv_order_cnt); break;
+                    case dataType_t::recv_exec_cnt : data.emplace_back((long long)stats.recv_exec_cnt); break;
+                    case dataType_t::recv_unknown_cnt : data.emplace_back((long long)stats.recv_unknown_cnt); break;
                     case dataType_t::last_frame_time_ms : data.emplace_back(utc_to_str(stats.last_frame_time_ms).c_str()); break;
-                    case dataType_t::live_code_nb : data.emplace_back(stats.live_code_nb); break;
-                    case dataType_t::live_instrument_snap_cnt : data.emplace_back(stats.live_instrument_snap_cnt); break;
-                    case dataType_t::live_index_snap_cnt : data.emplace_back(stats.live_index_snap_cnt); break;
-                    case dataType_t::live_order_cnt : data.emplace_back(stats.live_order_cnt); break;
-                    case dataType_t::live_exec_cnt : data.emplace_back(stats.live_exec_cnt); break;
+                    case dataType_t::live_code_nb : data.emplace_back((long long)stats.live_code_nb); break;
+                    case dataType_t::live_instrument_snap_cnt : data.emplace_back((long long)stats.live_instrument_snap_cnt); break;
+                    case dataType_t::live_index_snap_cnt : data.emplace_back((long long)stats.live_index_snap_cnt); break;
+                    case dataType_t::live_order_cnt : data.emplace_back((long long)stats.live_order_cnt); break;
+                    case dataType_t::live_exec_cnt : data.emplace_back((long long)stats.live_exec_cnt); break;
                 };
             }
